Changing composer install and a PHP setting.
Sadly the composer package in Ubuntu is very out of date even with the
latest PHP package repo so using the install script provided by
composer.

Changed the assert setting for PHP to match what it should be in
development while the package default configuration is set for
production. This was stopping the Drupal unit tests from finishing.